Create new dbd module
This commit introduces a new slurm dbd module, this is used to create a
separate slurmdbd instance, that instance is used to run slurmdbd, this
way multiple clusters can connect to this same dbd instance thus
enabling slurm multi-cluster.
In order to make this work all the clusters and the dbd instance need to
share the same munge and jwt key, to facilitate this, these key are
created in the dbd module and stored in google secret manager, to be
retrieved by the instances, this way a safe and reliable way to share
the keys is established.