Add hybrid setup function
The hybrid flag gets used to know whether to execute the regular main
function or the setup_hybrid one, the config retrieval is done by both,
so it has been wrote into it's own function.

Also the bucket flag has been added to specify the bucket where the
config is stored. This bucket is passed as parameter to the config
retrieval functions.

In the setup_hybrid, the only needed thing to do is to generate the
controller config files.

Some modifications are also added in order to not query the google
metadata server when the setup.py is run in the hybrid setup mode.