Change bin and log directory in the cloud
Adds the bool parameter use_same_path_on_cloud in hybrid_conf, if not
specified or explicitly set to false in the cloud it will use the
default slurm paths for bin and log, else it will use the paths
specified in slurm_log_dir and slurm_bin_dir. On premise it will always
use the slurm_bin_dir and slurm_log_dir.