Add munge secret variable
This variable in the hybrid_conf specified the name of a GCP secret that
contains the base64 encoded munge.key, to be used for the compute nodes
to download the munge key. This commit also removed the variable
munge_mount as it is not needed when using this munge_secret. The
regular clusters will still be using the munge_mount internally.